Allow two new parameters to be sent to comforter: mergeBase and targetBranch
Comforter uses master as the benchmark branch to compare the converage changes.
targetBranch allows you to specify the branch instead. If not included, it defaults to master
mergeBase allows you to specify a commit hash to use instead. If not include
    or coverage has no data for the commit, it defaults to targetBranch.

Note: this requires coverage-cli to be updated to a 4.4 version that passes along
    mergeBase and targetBranch